pyproject.toml
[project]
name = "spend-categorization"
version = "0.1.0"
description = "Invoice classification correction tool with Databricks integration"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "databricks-sql-connector>=3.0.0",
    "pandas>=2.0.0",
    "pyarrow>=14.0.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-mock>=3.12.0",
    "ipykernel>=6.25.0",
    "streamlit>=1.28.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/invoice_app"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

